+++ gremlin/src/test/java/org/apache/tinkerpop/gremlin/arcadedb/GremlinTest.java
@@ -67,7 +67,7 @@ public class GremlinTest {
         Assertions.assertTrue(row.getProperty("p.age") instanceof Number);
         Assertions.assertTrue((int) row.getProperty("p.age") > lastAge);
 
-        lastAge = (int) row.getProperty("p.age");
+        lastAge = row.getProperty("p.age");
       }
 
       Assertions.assertEquals(25, i);
@@ -105,7 +105,7 @@ public class GremlinTest {
         Assertions.assertTrue(row.getProperty("p.age") instanceof Number);
         Assertions.assertTrue((int) row.getProperty("p.age") > lastAge);
 
-        lastAge = (int) row.getProperty("p.age");
+        lastAge = row.getProperty("p.age");
       }
 
       Assertions.assertEquals(25, i);
