+++ gremlin/src/test/java/org/apache/tinkerpop/gremlin/arcadedb/SQLTest.java
@@ -62,7 +62,7 @@ public class SQLTest {
         Assertions.assertTrue(row.getProperty("p.age") instanceof Number);
         Assertions.assertTrue((int) row.getProperty("p.age") > lastAge);
 
-        lastAge = (int) row.getProperty("p.age");
+        lastAge = row.getProperty("p.age");
       }
 
       Assertions.assertEquals(25, i);
